requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy==1.18.5
pandas==1.1.5
pillow==7.2.0
tqdm